requirements.txt
APScheduler~=4.0.0a5
PyJWT~=2.10.1
PyYAML~=6.0.2
bcrypt==4.0.1
common~=0.1.2
dotenv~=0.9.9
httpx==1.0.0b0
infrastructure~=3.5.6
logging~=0.4.9.6
motor==3.7.0
pandas~=2.2.3
passlib~=1.7.4
phonenumbers~=9.0.1
pydantic-settings~=2.8.1
pydantic~=2.9.2
pymongo~=4.12.0
pytest-asyncio~=0.26.0
pytest~=8.3.5
python-dotenv~=1.1.0
python-jose~=3.4.0
redis~=6.0.0b1
requests~=2.32.3
slowapi~=0.1.9
starlette~=0.46.1
user-agents~=2.2.0
uvicorn~=0.34.0
celery~=5.5.0
aiohttp~=3.11.16
sentry-sdk~=2.25.1
fastapi~=0.115.12
tenacity~=8.5.0
colorama==0.4.6